[Refactoring] cleanup mailbox-adapter and mpt-imapmailbox-core modules from mailbox-store dependency
