+++ server/container/mailbox-adapter/src/main/java/org/apache/james/adapter/mailbox/UserRepositoryAuthenticator.java
@@ -17,40 +17,35 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.adapter.mailbox.store;
+package org.apache.james.adapter.mailbox;
 
 import javax.inject.Inject;
 
 import org.apache.james.core.Username;
-import org.apache.james.mailbox.Authorizator;
+import org.apache.james.mailbox.Authenticator;
 import org.apache.james.mailbox.exception.MailboxException;
 import org.apache.james.user.api.UsersRepository;
 import org.apache.james.user.api.UsersRepositoryException;
 
 /**
- * Authorizator which use an UsersRepository to check if the delegation is allowed
+ * Authenticator which use an UsersRepository to check if the user and password
+ * match
  */
-public class UserRepositoryAuthorizator implements Authorizator {
+public class UserRepositoryAuthenticator implements Authenticator {
 
     private final UsersRepository repos;
 
     @Inject
-    public UserRepositoryAuthorizator(UsersRepository repos) {
+    public UserRepositoryAuthenticator(UsersRepository repos) {
         this.repos = repos;
     }
 
     @Override
-    public AuthorizationState canLoginAsOtherUser(Username userId, Username otherUserId) throws MailboxException {
+    public boolean isAuthentic(Username userid, CharSequence passwd) throws MailboxException {
         try {
-            if (!repos.isAdministrator(userId)) {
-                return AuthorizationState.NOT_ADMIN;
-            }
-            if (!repos.contains(otherUserId)) {
-                return AuthorizationState.UNKNOWN_USER;
-            }
-            return AuthorizationState.ALLOWED;
+            return repos.test(userid, passwd.toString());
         } catch (UsersRepositoryException e) {
-            throw new MailboxException("Unable to access usersRepository", e);
+            throw new MailboxException("Unable to access UsersRepository", e);
         }
     }
