+++ server/container/mailbox-adapter/src/main/java/org/apache/james/adapter/mailbox/UserRepositoryAuthorizator.java
@@ -17,35 +17,40 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.adapter.mailbox.store;
+package org.apache.james.adapter.mailbox;
 
 import javax.inject.Inject;
 
 import org.apache.james.core.Username;
-import org.apache.james.mailbox.Authenticator;
+import org.apache.james.mailbox.Authorizator;
 import org.apache.james.mailbox.exception.MailboxException;
 import org.apache.james.user.api.UsersRepository;
 import org.apache.james.user.api.UsersRepositoryException;
 
 /**
- * Authenticator which use an UsersRepository to check if the user and password
- * match
+ * Authorizator which use an UsersRepository to check if the delegation is allowed
  */
-public class UserRepositoryAuthenticator implements Authenticator {
+public class UserRepositoryAuthorizator implements Authorizator {
 
     private final UsersRepository repos;
 
     @Inject
-    public UserRepositoryAuthenticator(UsersRepository repos) {
+    public UserRepositoryAuthorizator(UsersRepository repos) {
         this.repos = repos;
     }
 
     @Override
-    public boolean isAuthentic(Username userid, CharSequence passwd) throws MailboxException {
+    public AuthorizationState canLoginAsOtherUser(Username userId, Username otherUserId) throws MailboxException {
         try {
-            return repos.test(userid, passwd.toString());
+            if (!repos.isAdministrator(userId)) {
+                return AuthorizationState.NOT_ADMIN;
+            }
+            if (!repos.contains(otherUserId)) {
+                return AuthorizationState.UNKNOWN_USER;
+            }
+            return AuthorizationState.ALLOWED;
         } catch (UsersRepositoryException e) {
-            throw new MailboxException("Unable to access UsersRepository", e);
+            throw new MailboxException("Unable to access usersRepository", e);
         }
     }
