+++ server/container/spring/src/main/resources/META-INF/org/apache/james/spring-mailbox-authenticator.xml
@@ -27,6 +27,6 @@
       Mailbox Authenticator
      -->
 
-    <bean id="authenticator" class="org.apache.james.adapter.mailbox.store.UserRepositoryAuthenticator"/>
+    <bean id="authenticator" class="org.apache.james.adapter.mailbox.UserRepositoryAuthenticator"/>
 
 </beans>
