+++ server/container/spring/src/main/resources/META-INF/org/apache/james/spring-mailbox-authorizator.xml
@@ -27,6 +27,6 @@
       Mailbox Authorizator
      -->
 
-    <bean id="authorizator" class="org.apache.james.adapter.mailbox.store.UserRepositoryAuthorizator"/>
+    <bean id="authorizator" class="org.apache.james.adapter.mailbox.UserRepositoryAuthorizator"/>
 
 </beans>
